feat: preferredMailSender 1/2 - preferredMailSender system variable
This new preferredMailSender system variable defaults to
'SMTP|AmazonSES|SendGrid" which means it will first look for an SMTP
setting when sending email directly from Aspen.
If that fails or doesn't exist, it'll then query Amazon SES, and finally
SendGrid.
This order may be configured through the system variables UI.